src/api/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
autopep8==1.3.4
bcrypt==3.1.4
billiard==3.5.0.3
boto==2.39.0
boto3==1.6.9
botocore==1.9.9
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.1.4
defusedxml==0.4.1
Django==1.11.9
django-ajax-selects==1.7.0
django-cachalot==1.2.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.2.0
django-crispy-forms==1.6.0
django-filter==1.0.2
django-queryinspect==0.1.0
django-redis==4.3.0
django-rest-swagger==2.1.0
django-scarface==3.0.2
django-ses==0.8.5
django-silk==2.0.0
django-storages==1.6.5
django-suit==0.2.26
djangorestframework==3.5.3
djangorestframework-filters==0.10.2
djangorestframework-gis==0.12
djangorestframework-xml==1.3.0
django-sql-explorer==1.1.1
docutils==0.14
ecdsa==0.13
Fabric==1.10.2
Fabric3==1.14.post1
factory-boy==2.10.0
Faker==0.8.11
fcm-django==0.2.19
filechunkio==1.6
flake8==3.5.0
future==0.16.0
gprof2dot==2016.10.13
gunicorn==19.4.5
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
paramiko==2.4.0
pep8==1.7.1
Pillow==3.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
Pygments==2.2.0
PyNaCl==1.2.1
python-dateutil==2.6.1
python-memcached==1.57
pytz==2018.3
PyYAML==3.11
raven==6.5.0
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
simplejson==3.13.2
six==1.10.0
sqlparse==0.2.4
text-unidecode==1.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==13.1.2
django-debug-toolbar==1.9.1
arrow==0.12.1
django-pgviews==0.5.3
django-templated-mail==1.1.1
